tool that I used in order to register a car
The queue to register a car in my region is very long. The webpage enables to register every day for a day after 2 weeks. The problem is, that there is so many people willing to do it , that after 2 min there is no free spot. The other problem is that the time of appearance is not fixed. That's why I created a tool that based on the class from day from calendar that checks when it's possible to register and beeps. It requires a login to enter that page.